added value change BluetoothEvent
moved BluetoothEvent to a separate module

fixed examples/
