+++ src/bluetooth_session.rs
@@ -1,97 +1,12 @@
-use dbus::{
-    arg::cast, arg::RefArg, arg::TypeMismatchError, arg::Variant, BusType, ConnMsgs, Connection,
-    Message,
-};
-use std::collections::HashMap;
+use dbus::{BusType, ConnMsgs, Connection};
+
 use std::error::Error;
 
 static BLUEZ_MATCH: &'static str = "type='signal',sender='org.bluez'";
 
+#[derive(Debug)]
 pub struct BluetoothSession {
     connection: Connection,
-}
-
-#[derive(Clone)]
-pub enum BluetoothEvent {
-    Powered {
-        object_path: String,
-        powered: bool,
-    },
-    Discovering {
-        object_path: String,
-        discovering: bool,
-    },
-    Connected {
-        object_path: String,
-        connected: bool,
-    },
-    ServicesResolved {
-        object_path: String,
-        services_resolved: bool,
-    },
-    None,
-}
-
-impl BluetoothEvent {
-    pub fn from(conn_msg: Message) -> Option<BluetoothEvent> {
-        let result: Result<
-            (&str, HashMap<String, Variant<Box<RefArg>>>),
-            TypeMismatchError,
-        > = conn_msg.read2();
-
-        match result {
-            Ok((_, properties)) => {
-                let object_path = conn_msg.path().unwrap().to_string();
-
-                if let Some(value) = properties.get("Powered") {
-                    if let Some(powered) = cast::<bool>(&value.0) {
-                        let event = BluetoothEvent::Powered {
-                            object_path: object_path.clone(),
-                            powered: *powered,
-                        };
-
-                        return Some(event);
-                    }
-                }
-
-                if let Some(value) = properties.get("Discovering") {
-                    if let Some(discovering) = cast::<bool>(&value.0) {
-                        let event = BluetoothEvent::Discovering {
-                            object_path: object_path.clone(),
-                            discovering: *discovering,
-                        };
-
-                        return Some(event);
-                    }
-                }
-
-                if let Some(value) = properties.get("Connected") {
-                    if let Some(connected) = cast::<bool>(&value.0) {
-                        let event = BluetoothEvent::Connected {
-                            object_path: object_path.clone(),
-                            connected: *connected,
-                        };
-
-                        return Some(event);
-                    }
-                }
-
-                if let Some(value) = properties.get("ServicesResolved") {
-                    if let Some(services_resolved) = cast::<bool>(&value.0) {
-                        let event = BluetoothEvent::ServicesResolved {
-                            object_path: object_path.clone(),
-                            services_resolved: *services_resolved,
-                        };
-
-                        return Some(event);
-                    }
-                }
-
-                Some(BluetoothEvent::None)
-            }
-            Err(err) => None,
-        }
-    }
 }
 
 impl BluetoothSession {
@@ -123,5 +38,3 @@ impl BluetoothSession {
         self.connection.incoming(timeout_ms)
     }
 }
-
-unsafe impl Send for BluetoothEvent {}
